+++ .gitignore
@@ -8,3 +8,4 @@ __pycache__
 dist/
 build/
 *.egg-info
+examples/*/*.html
